# reduce order size - performance improvement