--- a/src/nasdaq/nordic_itch_handler.cc
+++ b/src/nasdaq/nordic_itch_handler.cc
@@ -124,7 +124,7 @@ void nordic_itch_handler::process_msg(const itch_add_order* m)
 
         uint64_t order_id = itch_uatoi(m->OrderReferenceNumber, sizeof(m->OrderReferenceNumber));
         uint64_t price    = itch_uatoi(m->Price, sizeof(m->Price));;
-        uint64_t quantity = itch_uatoi(m->Quantity, sizeof(m->Quantity));;
+        uint32_t quantity = itch_uatoi(m->Quantity, sizeof(m->Quantity));;
         auto     side     = itch_side(m->BuySellIndicator);
 
         order o{order_id, price, quantity, side, timestamp()};
@@ -145,7 +145,7 @@ void nordic_itch_handler::process_msg(const itch_add_order_mpid* m)
 
         uint64_t order_id = itch_uatoi(m->OrderReferenceNumber, sizeof(m->OrderReferenceNumber));
         uint64_t price    = itch_uatoi(m->Price, sizeof(m->Price));;
-        uint64_t quantity = itch_uatoi(m->Quantity, sizeof(m->Quantity));;
+        uint32_t quantity = itch_uatoi(m->Quantity, sizeof(m->Quantity));;
         auto     side     = itch_side(m->BuySellIndicator);
 
         order o{order_id, price, quantity, side, timestamp()};