before
[denis@f22 helix]$ ./build/tests/order_book_perf_test 
order_book::add()    66 ns/op
order_book::cancel() 11 ns/op
order_book::remove() 34 ns/op

after
[denis@f22 helix]$ ./build/tests/order_book_perf_test 
order_book::add()    59 ns/op
order_book::cancel() 10 ns/op
order_book::remove() 31 ns/op